Use resourceful route for message reply
